+++ apps/server/src/shared/common/error/business.error.spec.ts
