--- apps/server/src/shared/error/index.ts
@@ -1,2 +1,0 @@
-export * from './business.error';
-export * from './api-validation.error';
